Added ParsedQueryTestHelpers to get rid of code dupplication.
